Tests/LibWeb/Text/expected/Editing/execCommand-preserveWhitespace.txt

Before: foo<div style="white-space: pre">bar</div>
After: foo<span style="white-space: pre;">bar</span>
